test: split up tests, remove module mock
Also do not reuse static objects
