+++ historicalData/historicalDataRepository.ts
@@ -11,9 +11,10 @@ import {
 } from '@hello.nrfcloud.com/proto/hello/history'
 import { parseResult } from '@nordicsemiconductor/timestream-helpers'
 import { type Static } from '@sinclair/typebox'
-import { groupBy } from 'lodash-es'
-import { getQueryStatement } from './queryGenerator.js'
+import { getQueryStatement } from './getQueryStatement.js'
 import { createTrailOfCoordinates } from '../util/createTrailOfCoordinates.js'
+import { transformTimestreamData } from './transformTimestreamData.js'
+import { normalizedData } from './normalizedData.js'
 
 export const HistoricalChartTypes = {
 	lastHour: {
@@ -41,62 +42,6 @@ export const HistoricalChartTypes = {
 
 export type HistoricalRequest = Static<typeof HistoricalDataRequest>
 export type HistoricalResponse = Static<typeof HistoricalDataResponse>
-
-export const normalizedData = (
-	data: Record<string, unknown>[],
-): Record<string, unknown>[] =>
-	data.map((o) => {
-		const key = 'measure_name' in o ? (o['measure_name'] as string) : undefined
-		const val =
-			'measure_value::double' in o ? o['measure_value::double'] : undefined
-		if (key !== undefined && val !== undefined) {
-			o[key] = val
-		}
-
-		return o
-	})
-
-export const transformTimestreamData = (
-	data: Record<string, unknown>[],
-	mapKeys: { fromKey: string; toKey: string }[],
-): Record<string, unknown>[] => {
-	const transformedData = []
-
-	// Added ts to mapKeys as default
-	mapKeys.push({ fromKey: 'time', toKey: 'ts' })
-
-	const groupedData = groupBy(data, (d) => d.time)
-	for (const item in groupedData) {
-		const data = groupedData[item]
-		const transformedRecord = data?.reduce<Record<string, unknown>>(
-			(result, record) => {
-				result = {
-					...mapKeys.reduce<Record<string, unknown>>(
-						(resultMapKeys, mapKey) => {
-							if (mapKey.fromKey in record) {
-								if (mapKey.fromKey === 'time') {
-									resultMapKeys[mapKey.toKey] = (
-										record[mapKey.fromKey] as Date
-									).getTime()
-								} else {
-									resultMapKeys[mapKey.toKey] = record[mapKey.fromKey]
-								}
-							}
-							return resultMapKeys
-						},
-						{},
-					),
-					...result,
-				}
-				return result
-			},
-			{},
-		)
-		if (transformedRecord !== undefined) transformedData.push(transformedRecord)
-	}
-
-	return transformedData
-}
 
 export const historicalDataRepository = ({
 	timestream,
